+++ BlowinCleanCode/BlowinCleanCode.Test/Verifiers/CSharpAnalyzerVerifier`1.cs
@@ -1,10 +1,14 @@
-﻿using System.Threading;
+﻿using System;
+using System.Collections;
+using System.Runtime.Serialization;
+using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Testing;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Testing.Verifiers;
+using Xunit;
 
 namespace BlowinCleanCode.Test.Verifiers
 {
@@ -32,7 +36,44 @@ namespace BlowinCleanCode.Test.Verifiers
             };
 
             test.ExpectedDiagnostics.AddRange(expected);
-            await test.RunAsync(CancellationToken.None);
+            try
+            {
+                await test.RunAsync(CancellationToken.None);
+            }
+            catch(Exception e)
+            {
+                throw new WithSourceMessageException(source, e);
+            }
+        }
+        
+        private sealed class WithSourceMessageException : Exception
+        {
+            private string _source;
+            private Exception _ex;
+
+            public override string Message => _source + Environment.NewLine + _ex.Message;
+
+            public override IDictionary Data => _ex.Data;
+
+            public override string Source => _ex.Source;
+
+            public override string HelpLink
+            {
+                get => _ex.HelpLink;
+                set => _ex.HelpLink = value;
+            }
+
+            public override string StackTrace => _ex.StackTrace;
+
+            public WithSourceMessageException(string source, Exception ex)
+            {
+                _source = source;
+                _ex = ex;
+            }
+
+            public override Exception GetBaseException() => _ex.GetBaseException();
+
+            public override void GetObjectData(SerializationInfo info, StreamingContext context) => _ex.GetObjectData(info, context);
         }
     }
 }